Implemented generic saving, loading in IFileservice
