+++ NOTEA/Conspects/Fizika.txt
@@ -1,1 +1,1 @@
-{"Date":"2023-10-04","Name":"Fizika","ConspectText":"Pirmasis Niutono dėsnis: egzistuoja tokios atskaitos sistemos, kuriose kūnas yra rimties būsenos arba juda tolygiai ir tiesiaeigiškai, jei jį veikiančių jėgų dedamoji lygi nuliui.\r\n\r\nAntrasis Niutono dėsnis: kūno įgyjamas pagreitis (a), yra tiesiogiai proporcingas kūną veikiančių jėgų dedamajai ir atvirkščiai proporcingas kūno masei. a=F/m\r\n\r\nTrečiasis Niutono dėsnis: dviejų kūnų sąveikos jėgos yra lygaus dydžio ir priešingų krypčių. F1=-F2"}
+{"Date":"2023-10-09","ConspectType":0,"Name":"Fizika","ConspectText":"Pirmasis Niutono dėsnis: egzistuoja tokios atskaitos sistemos, kuriose kūnas yra rimties būsenos arba juda tolygiai ir tiesiaeigiškai, jei jį veikiančių jėgų dedamoji lygi nuliui.\r\n\r\nAntrasis Niutono dėsnis: kūno įgyjamas pagreitis (a), yra tiesiogiai proporcingas kūną veikiančių jėgų dedamajai ir atvirkščiai proporcingas kūno masei. a=F/m\r\n\r\nTrečiasis Niutono dėsnis: dviejų kūnų sąveikos jėgos yra lygaus dydžio ir priešingų krypčių. F1=-F2"}
