+++ NOTEA/Conspects/PSI events.txt
@@ -1,0 +1,1 @@
+{"Date":"2023-10-09","ConspectType":2,"Name":"PSI events","ConspectText":"Events can be raised only from within the class that has declared it."}
