+++ NOTEA/Models/IConspectModel.cs
@@ -1,0 +1,7 @@
+﻿namespace NOTEA.Models
+{
+    public interface IConspectModel
+    {
+        string Name { get; }
+    }
+}
