--- NOTEA/Models/IDisplayable.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.CodeAnalysis.CSharp.Syntax;
-
-namespace NOTEA.Models
-{
-    public interface IDisplayable <DataType>
-    {
-        public DataType Display();
-    }
-}
