+++ NOTEA/Views/Conspect/ConspectList.cshtml
@@ -1,4 +1,4 @@
-﻿@model NOTEA.Models.ConspectListModel
+﻿@model NOTEA.Models.ConspectListModel<ConspectModel>
  
 @{
     ViewData["Title"] = "Conspect list";
